Updated timeline view to automatically update when a new behavior is added to it.
