+++ .idea/modules.xml
@@ -2,7 +2,7 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
-      <module fileurl="file://$PROJECT_DIR$/Sequencer.iml" filepath="$PROJECT_DIR$/Sequencer.iml" />
+      <module fileurl="file://$PROJECT_DIR$/Protocolizer.iml" filepath="$PROJECT_DIR$/Protocolizer.iml" />
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
     </modules>
   </component>
